\name{Data_Soybean}
\alias{Data_Soybean}
\docType{data}
\title{
Monthly International Soybean Oil Price
}
\description{
Monthly international Soyabean oil price from August 2001 to June 2020.
}
\usage{data("Data_Soybean")}
\format{
  A time series data with 227 observations.
  \describe{
    \item{\code{price}}{a time series}
  }
}
\details{
Dataset contains 227 observations of monthly international Soyabean oil price. It is obtained from World Bank "Pink sheet".
}
\source{
https://www.worldbank.org/en/research/commodity-markets
}
\references{
https://www.worldbank.org/en/research/commodity-markets
}
\examples{
data(Data_Soybean)
}
\keyword{datasets}
